Backend > Structures (models, repositories, dtos) and AuthController -- Register and Login functionalities (#20)
Created most of the backend structures and implemented register and login endpoint logic.

if running backend service locally, currently h2 database temporary instance is spun up and it is possible to 

register a new user  

and login (obtain token for) with existent user

Error handling is implemented in case attempting to register with existing username, or if sending request with missing or null fields.


* initial config dependencies

* Initial structures for Auth, User, Group, and Matches. Preparing to make calls with H2 and postman

* Successfully implemented register and login endpoints + some refactors and fixes on structures

* Enhanced error handling

* quick updated on register return status code to 201 CREATED